Apply PIMPL idiom to sort-merge join (#21349)
Depends on https://github.com/rapidsai/cudf/pull/21314
Follow-up work from discussions in [comment](https://github.com/rapidsai/cudf/pull/21314#pullrequestreview-3757423163)

- Apply the pimpl (pointer to implementation) idiom to `sort_merge_join`, following the pattern established in `hash_join`
- Remove deprecated `sort_merge_inner_join` and `merge_inner_join` APIs

## Changes
- Create new detail header `include/cudf/detail/join/sort_merge_join.hpp` containing the full implementation class
- Refactor public header `include/cudf/join/sort_merge_join.hpp` to a minimal wrapper with `std::unique_ptr<impl_type const> _impl`
- Move implementation methods into `namespace detail { }` block in `sort_merge_join.cu`
- Add wrapper class methods that delegate to `_impl`

## Benefits
- Hides implementation details (`preprocessed_table`, `sort_merge_join_match_context`, internal methods) from the public API
- Reduces compilation dependencies for consumers of the public header
- Improves API stability - implementation changes no longer affect the public header
- Consistent design pattern with `hash_join`

## Test plan
- Existing join tests should pass unchanged (API is compatible)

Authors:
  - Shruti Shivakumar (https://github.com/shrshi)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Tianyu Liu (https://github.com/kingcrimsontianyu)

URL: https://github.com/rapidsai/cudf/pull/21349
