+++ cpp/include/cudf/join/sort_merge_join.hpp
@@ -5,8 +5,6 @@
 
 #pragma once
 
-#include <cudf/column/column.hpp>
-#include <cudf/column/column_view.hpp>
 #include <cudf/join/join.hpp>
 #include <cudf/table/table_view.hpp>
 #include <cudf/types.hpp>
@@ -14,8 +12,9 @@
 #include <cudf/utilities/memory_resource.hpp>
 
 #include <rmm/cuda_stream_view.hpp>
+#include <rmm/device_uvector.hpp>
 
-#include <optional>
+#include <memory>
 
 namespace CUDF_EXPORT cudf {
 
@@ -24,6 +23,11 @@ namespace CUDF_EXPORT cudf {
  * @{
  * @file
  */
+
+// Forward declaration
+namespace detail {
+class sort_merge_join;
+}
 
 /**
  * @brief Class that implements sort-merge algorithm for table joins
@@ -35,7 +39,8 @@ namespace CUDF_EXPORT cudf {
  */
 class sort_merge_join {
  public:
-  sort_merge_join()                                  = delete;
+  sort_merge_join() = delete;
+  ~sort_merge_join();
   sort_merge_join(sort_merge_join const&)            = delete;
   sort_merge_join(sort_merge_join&&)                 = delete;
   sort_merge_join& operator=(sort_merge_join const&) = delete;
@@ -197,244 +202,9 @@ class sort_merge_join {
     rmm::device_async_resource_ref mr = cudf::get_current_device_resource_ref()) const;
 
  private:
-  /**
-   * @brief Helper struct to pre-process tables before join operations
-   */
-  struct preprocessed_table {
-    table_view _table_view;  ///< Unprocessed table view before pre-processing
-
-    table_view
-      _null_processed_table_view;  ///< Processed table view which is the null-free subset of the
-                                   ///< rows of the unprocessed table view if null equality is set
-                                   ///< to false, otherwise equal to the unprocessed table view
-
-    std::optional<rmm::device_buffer> _validity_mask =
-      std::nullopt;  ///< Optional validity mask for null_equality::UNEQUAL case
-    std::optional<size_type> _num_nulls =
-      std::nullopt;  ///< Optional count of nulls for null_equality::UNEQUAL case
-    std::optional<std::unique_ptr<table>> _null_processed_table =
-      std::nullopt;  ///< Optional filtered table for null_equality::UNEQUAL case
-
-    std::optional<std::unique_ptr<column>> _null_processed_table_sorted_order =
-      std::nullopt;  ///< Optional sort ordering for pre-sorted tables
-
-    /**
-     * @brief Factory method to create a preprocessed table
-     *
-     * @param table The table to preprocess
-     * @param compare_nulls Controls whether null join-key values should match or not
-     * @param is_sorted Enum to indicate if the table is pre-sorted
-     * @param stream CUDA stream used for device memory operations and kernel launches
-     * @return A preprocessed_table ready for join operations
-     */
-    static preprocessed_table create(table_view const& table,
-                                     null_equality compare_nulls,
-                                     sorted is_sorted,
-                                     rmm::cuda_stream_view stream);
-
-    /**
-     * @brief Mark rows in unprocessed table with nulls at root or child levels by populating the
-     * _validity_mask
-     *
-     * @param stream CUDA stream used for device memory operations and kernel launches
-     */
-    void populate_nonnull_filter(rmm::cuda_stream_view stream);
-
-    /**
-     * @brief Apply _validity_mask to the _table_view to create a null-free table
-     *
-     * @param stream CUDA stream used for device memory operations and kernel launches
-     */
-    void apply_nonnull_filter(rmm::cuda_stream_view stream);
-
-    /**
-     * @brief Pre-process the unprocessed table when null equality is set to unequal
-     *
-     * @param stream CUDA stream used for device memory operations and kernel launches
-     */
-    void preprocess_unprocessed_table(rmm::cuda_stream_view stream);
-
-    /**
-     * @brief Compute sorted ordering of the processed table
-     *
-     * @param stream CUDA stream used for device memory operations and kernel launches
-     */
-    void compute_sorted_order(rmm::cuda_stream_view stream);
-
-    /**
-     * @brief Create mapping from processed table indices to unprocessed table indices
-     *
-     * @param stream CUDA stream used for device memory operations and kernel launches
-     * @return A device vector containing the mapping from processed table indices to unprocessed
-     * table indices
-     */
-    rmm::device_uvector<size_type> map_table_to_unprocessed(rmm::cuda_stream_view stream) const;
-  };
-
-  /**
-   * @brief Extended match context for sort-merge join with preprocessed table state.
-   *
-   * This struct derives from join_match_context and adds the preprocessed left table
-   * state needed for thread-safe partitioned join operations. It is returned by
-   * inner_join_match_context() and used with partitioned_inner_join().
-   */
-  struct sort_merge_join_match_context : public join_match_context {
-    preprocessed_table preprocessed_left;  ///< Preprocessed left table state for partitioned joins
-
-    /**
-     * @brief Construct a sort_merge_join_match_context
-     *
-     * @param left_table The left table view
-     * @param match_counts Device vector of match counts per row
-     * @param preprocessed Preprocessed left table state
-     */
-    sort_merge_join_match_context(table_view left_table,
-                                  std::unique_ptr<rmm::device_uvector<size_type>> match_counts,
-                                  preprocessed_table preprocessed)
-      : join_match_context{left_table, std::move(match_counts)},
-        preprocessed_left{std::move(preprocessed)}
-    {
-    }
-  };
-
-  preprocessed_table preprocessed_right;  ///< Preprocessed right table
-  null_equality compare_nulls;            ///< Null comparison mode
-
-  /**
-   * @brief Post-process left and right tables after the merge operation
-   *
-   * @param preprocessed_left The preprocessed left table
-   * @param smaller_indices Indices for the smaller processed table used to construct the join
-   * result
-   * @param larger_indices Indices for the larger processed table used to construct the join result
-   * @param stream CUDA stream used for device memory operations and kernel launches
-   */
-  void postprocess_indices(preprocessed_table const& preprocessed_left,
-                           device_span<size_type> smaller_indices,
-                           device_span<size_type> larger_indices,
-                           rmm::cuda_stream_view stream) const;
-
-  /**
-   * @brief Core merge operation implementation for the sort-merge join algorithm.
-   *
-   * This template method performs the actual merge operation between preprocessed left and
-   * right tables for different types of joins. It serves as the common implementation for
-   * various join methods (inner_join, inner_join_match_context, partitioned_inner_join).
-   *
-   * The method takes a generic merge operation functor that defines the specific join
-   * behavior to be applied during the merge phase. This design allows for different join
-   * operations (such as generating indices or counting matches) to share the same
-   * underlying merge algorithm.
-   *
-   * The method expects that tables have already been preprocessed to handle
-   * null values according to the null_equality setting.
-   *
-   * @tparam MergeOperation Functor type that implements the specific join operation
-   * @param preprocessed_left The preprocessed left table
-   * @param right_view The preprocessed right table view
-   * @param left_view The preprocessed left table view
-   * @param op The merge operation functor to execute during the merge
-   * @param stream CUDA stream used for device memory operations and kernel launches
-   *
-   * @return The result of the merge operation as defined by the MergeOperation functor
-   *         (typically pairs of join indices or match counts)
-   */
-  template <typename MergeOperation>
-  auto invoke_merge(preprocessed_table const& preprocessed_left,
-                    table_view right_view,
-                    table_view left_view,
-                    MergeOperation&& op,
-                    rmm::cuda_stream_view stream) const;
+  using impl_type = cudf::detail::sort_merge_join;  ///< Implementation type
+  std::unique_ptr<impl_type const> _impl;           ///< Pointer to implementation
 };
-
-/**
- * @brief Returns a pair of row index vectors corresponding to an
- * inner join between the specified tables.
- *
- * The first returned vector contains the row indices from the left
- * table that have a match in the right table (in unspecified order).
- * The corresponding values in the second returned vector are
- * the matched row indices from the right table.
- *
- * @deprecated Use the object-oriented sort_merge_join API `cudf::sort_merge_join::inner_join`
- * instead
- *
- * @code{.pseudo}
- * Left: {{0, 1, 2}}
- * Right: {{1, 2, 3}}
- * Result: {{1, 2}, {0, 1}}
- *
- * Left: {{0, 1, 2}, {3, 4, 5}}
- * Right: {{1, 2, 3}, {4, 6, 7}}
- * Result: {{1}, {0}}
- * @endcode
- *
- * @throw std::invalid_argument if number of elements in `left_keys` or `right_keys`
- * mismatch.
- *
- * @param[in] left_keys The left table
- * @param[in] right_keys The right table
- * @param[in] compare_nulls controls whether null join-key values
- * should match or not.
- * @param stream CUDA stream used for device memory operations and kernel launches
- * @param mr Device memory resource used to allocate the returned table and columns' device memory
- *
- * @return A pair of vectors [`left_indices`, `right_indices`] that can be used to construct
- * the result of performing an inner join between two tables with `left_keys` and `right_keys`
- * as the join keys .
- */
-[[deprecated]] std::pair<std::unique_ptr<rmm::device_uvector<size_type>>,
-                         std::unique_ptr<rmm::device_uvector<size_type>>>
-sort_merge_inner_join(cudf::table_view const& left_keys,
-                      cudf::table_view const& right_keys,
-                      null_equality compare_nulls       = null_equality::EQUAL,
-                      rmm::cuda_stream_view stream      = cudf::get_default_stream(),
-                      rmm::device_async_resource_ref mr = cudf::get_current_device_resource_ref());
-
-/**
- * @brief Returns a pair of row index vectors corresponding to an inner join between the specified
- * tables.
- *
- * Assumes pre-sorted inputs and performs only the merge step.
- * The first returned vector contains the row indices from the left
- * table that have a match in the right table (in unspecified order).
- * The corresponding values in the second returned vector are
- * the matched row indices from the right table.
- *
- * @deprecated Use the object-oriented sort_merge_join API `cudf::sort_merge_join::inner_join`
- * instead
- *
- * @code{.pseudo}
- * Left: {{0, 1, 2}}
- * Right: {{1, 2, 3}}
- * Result: {{1, 2}, {0, 1}}
- *
- * Left: {{0, 1, 2}, {3, 4, 5}}
- * Right: {{1, 2, 3}, {4, 6, 7}}
- * Result: {{1}, {0}}
- * @endcode
- *
- * @throw std::invalid_argument if number of elements in `left_keys` or `right_keys`
- * mismatch.
- *
- * @param[in] left_keys The left table
- * @param[in] right_keys The right table
- * @param[in] compare_nulls controls whether null join-key values
- * should match or not.
- * @param stream CUDA stream used for device memory operations and kernel launches
- * @param mr Device memory resource used to allocate the returned table and columns' device memory
- *
- * @return A pair of vectors [`left_indices`, `right_indices`] that can be used to construct
- * the result of performing an inner join between two tables with `left_keys` and `right_keys`
- * as the join keys .
- */
-[[deprecated]] std::pair<std::unique_ptr<rmm::device_uvector<size_type>>,
-                         std::unique_ptr<rmm::device_uvector<size_type>>>
-merge_inner_join(cudf::table_view const& left_keys,
-                 cudf::table_view const& right_keys,
-                 null_equality compare_nulls       = null_equality::EQUAL,
-                 rmm::cuda_stream_view stream      = cudf::get_default_stream(),
-                 rmm::device_async_resource_ref mr = cudf::get_current_device_resource_ref());
 
 /** @} */  // end of group
 }  // namespace CUDF_EXPORT cudf
